Fixed broken Script::is_running method.

+++ libraries/Script.php
@@ -99,8 +99,6 @@ class Script extends Engine
 
     const DIR_LOCK = "/var/run/webconfig";
     const LOCK_SUFFIX = ".pid";
-    const CMD_PS = '/bin/ps';
-    const CMD_CAT = '/bin/cat';
 
     ///////////////////////////////////////////////////////////////////////////////
     // M E T H O D S
@@ -121,7 +119,8 @@ class Script extends Engine
         else
             $this->script_name = $script_name;
 
-        $this->lock_file = self::DIR_LOCK . '/' . $this->script_name . self::LOCK_SUFFIX;
+        $this->lock_file = sprintf('%s/%s%s',
+            self::DIR_LOCK, $this->script_name, self::LOCK_SUFFIX);
     }
 
     /**
@@ -137,32 +136,35 @@ class Script extends Engine
     public function lock($interval = 0, $retries = 0)
     {
         clearos_profile(__METHOD__, __LINE__);
+
         try {
-            $file = new File($this->lock_file, TRUE);
-            if ($file->exists()) {
-                $this->pid = $file->get_contents();
-                $running = $this->is_running();
-                if ($running) {
-                    for ($counter = 0; $counter < $retries; $counter++) {
-                        sleep($interval);
-                        $running = $this->is_running();
-                    }
-                    if ($running) {
-                        clearos_log($this->script_name, 'Unable to start script - currently running.');
-                        return FALSE;
-                    }
-                } else {
-                    clearos_log($this->script_name, 'Dead PID found - removing lock file.');
-                    $file->delete();
-                }
+            do {
+                if (! $this->is_running()) break;
+                sleep($interval);
+            } while (--$retries > 0);
+
+            if ($this->is_running()) {
+                clearos_log($this->script_name,
+                    'Unable to start script - currently running.');
+                return FALSE;
             }
+
+            $file = new File($this->lock_file, TRUE);
+            if ($file->exists()) $file->delete();
+
             $file->create('webconfig', 'webconfig', '0644');
             $this->pid = getmypid();
             $file->add_lines($this->pid);
+
             return TRUE;
+
         } catch (Exception $e) {
-            clearos_log($this->script_name, sprintf('Unable to create lock file: %s.', clearos_exception_message($e)));
-            throw new Engine_Exception(lang('base_unable_to_create_lock_file'), CLEAROS_WARNING);
+            clearos_log($this->script_name,
+                sprintf('Unable to create lock file: %s.',
+                    clearos_exception_message($e)));
+            throw new Engine_Exception(
+                lang('base_unable_to_create_lock_file'), CLEAROS_WARNING
+            );
         }
     }
 
@@ -176,12 +178,16 @@ class Script extends Engine
     public function unlock()
     {
         clearos_profile(__METHOD__, __LINE__);
+
         try {
+
             $file = new File($this->lock_file, TRUE);
-            if ($file->exists())
-                $file->delete();
+            if ($file->exists()) $file->delete();
+
         } catch (Exception $e) {
-            throw new Engine_Exception(lang('base_unable_to_remove_lock_file'), CLEAROS_WARNING);
+            throw new Engine_Exception(
+                lang('base_unable_to_remove_lock_file'), CLEAROS_WARNING
+            );
         }
     }
 
@@ -198,25 +204,29 @@ class Script extends Engine
 
         try {
 
-            $shell = new Shell();
-            $exitcode = $shell->execute(self::CMD_PS, " -e | awk '{print $1}'");
-            if ($exitcode != 0)
-                throw new Engine_Exception($shell->get_first_output_line(), CLEAROS_WARNING);
-            $pids = $shell->get_output();
-            if (in_array($this->pid, $pids)) {
-                // Check remote case where PID has been re-used by another process
-                $shell->execute(self::CMD_CAT, "/proc/" . $this->pid . '/cmdline');
-                $cmdline = trim($shell->get_last_output_line());
-                if (preg_match('/' . $this->script_name . "/", $cmdline))
-                    return TRUE;
-            }
-            return FALSE;
+            $file = new File($this->lock_file, TRUE);
+            if (!$file->exists()) return FALSE;
+            $this->pid = trim($file->get_contents());
+
+            $file = new File("/proc/{$this->pid}/cmdline", TRUE);
+            if (!$file->exists()) return FALSE;
+            $cmdline = trim($file->get_contents());
+
+            if (preg_match('/' . $this->script_name . "/", $cmdline))
+                return TRUE;
+
         } catch (Exception $e) {
-            clearos_log($this->script_name, sprintf('Unknown status: %s.', clearos_exception_message($e)));
+            clearos_log($this->script_name,
+                sprintf('Unknown status: %s.', clearos_exception_message($e)));
             throw new Engine_Exception(
-                lang('base_unknown_script_state') . ' (' . $this->script_name . ') - ' .
-                clearos_exception_message($e), CLEAROS_WARNING
+                lang('base_unknown_script_state') .
+                    ' (' . $this->script_name . ') - ' .
+                    clearos_exception_message($e), CLEAROS_WARNING
             );
         }
+
+        return FALSE;
     }
 }
+
+// vi: expandtab shiftwidth=4 softtabstop=4 tabstop=4
